+++ scratch.js
@@ -3,8 +3,8 @@ const readFile = require('./tests/lib/_cachedPage');
 // const wtf = require('./builds/wtf_wikipedia');
 // const wtf = require('./build');
 
-// var str = `{{Redirect|City of Toronto|the municipal government|Municipal government of Toronto|the historical part of the city prior to the 1998 amalgamation|Old Toronto}}`;
-var str = `{{tag|ref|content=haha}}`;
+var str = `{{Redirect|City of Toronto|the municipal government|Municipal government of Toronto|the historical part of the city prior to the 1998 amalgamation|Old Toronto}}`;
+// var str = `{{tag|ref|content=haha}}`;
 var doc = wtf(str);
 // console.log(doc.plaintext())
 console.log(doc.templates(0));
