+++ src/templates/index.js
@@ -1,6 +1,6 @@
 // const findRecursive = require('../../lib/recursive_match');
 const getName = require('./parsers/_getName');
-const getTemplates = require('./getTemplates');
+const getTemplates = require('./parsers/_getTemplates');
 
 const dates = require('./dates');
 const geo = require('./geo');
@@ -10,78 +10,42 @@ const generic = require('./generic');
 const links = require('./links');
 const formatting = require('./formatting');
 const interwiki = require('./interwiki');
-const ipa = require('./ipa');
-const ignore = require('./_ignore');
+const pronounce = require('./pronounce');
+const ignore = require('./ignore');
+
+//put them all together
+const inlineParsers = Object.assign({}, dates, inline, links, formatting, interwiki);
+const bigParsers = Object.assign({}, geo, pronounce, misc);
 
 const doTemplate = function(tmpl, wiki, r) {
   let name = getName(tmpl);
   //we explicitly ignore these templates
-  if (ignore.hasOwnProperty(name)) {
+  if (ignore.hasOwnProperty(name) === true) {
     wiki = wiki.replace(tmpl, '');
     return wiki;
   }
-  //date templates
-  if (dates.hasOwnProperty(name)) {
-    let str = dates[name](tmpl, r);
+  //string-replacement templates
+  if (inlineParsers.hasOwnProperty(name) === true) {
+    let str = inlineParsers[name](tmpl, r);
     wiki = wiki.replace(tmpl, str);
     return wiki;
   }
-  //geo templates
-  if (geo.hasOwnProperty(name)) {
-    let obj = geo[name](tmpl);
+  //section-template parsers
+  if (bigParsers.hasOwnProperty(name) === true) {
+    let obj = bigParsers[name](tmpl);
     if (obj) {
       r.templates.push(obj);
     }
     wiki = wiki.replace(tmpl, '');
     return wiki;
   }
-  if (ipa.hasOwnProperty(name)) {
-    let obj = ipa[name](tmpl);
-    if (obj) {
-      r.templates.push(obj);
-    }
-    wiki = wiki.replace(tmpl, '');
-    return wiki;
-  }
-  //inline templates
-  if (inline.hasOwnProperty(name)) {
-    let str = inline[name](tmpl);
-    wiki = wiki.replace(tmpl, str);
-    return wiki;
-  }
-  //inline templates
-  if (links.hasOwnProperty(name)) {
-    let str = links[name](tmpl);
-    wiki = wiki.replace(tmpl, str);
-    return wiki;
-  }
-  //inline formatting templates
-  if (formatting.hasOwnProperty(name)) {
-    let str = formatting[name](tmpl);
-    wiki = wiki.replace(tmpl, str);
-    return wiki;
-  }
-  if (interwiki.hasOwnProperty(name)) {
-    let str = interwiki[name](tmpl);
-    wiki = wiki.replace(tmpl, str);
-    return wiki;
-  }
-  //other ones
-  if (misc.hasOwnProperty(name)) {
-    let obj = misc[name](tmpl);
-    if (obj) {
-      r.templates.push(obj);
-    }
-    wiki = wiki.replace(tmpl, '');
-    return wiki;
-  }
+  //fallback parser
   let obj = generic(tmpl, name);
   if (obj) {
     r.templates.push(obj);
     wiki = wiki.replace(tmpl, '');
     return wiki;
   }
-
   //bury this template, if we don't know it
   // console.log(`  - no parser for '${name}' -`);
   // console.log('');
