--- src/templates/interwiki/sisters.js
@@ -1,18 +1,0 @@
-const sisterProjects = {
-  wikt: 'wiktionary',
-  commons: 'commons',
-  c: 'commons',
-  commonscat: 'commonscat',
-  n: 'wikinews',
-  q: 'wikiquote',
-  s: 'wikisource',
-  a: 'wikiauthor',
-  b: 'wikibooks',
-  voy: 'wikivoyage',
-  v: 'wikiversity',
-  d: 'wikidata',
-  species: 'wikispecies',
-  m: 'meta',
-  mw: 'mediawiki'
-};
-module.exports = sisterProjects;
