+++ src/templates/misc.js
@@ -1,9 +1,9 @@
 // const parseCitation = require('./citation');
 // const parseGeo = require('../geo');
-const keyValue = require('../parsers/keyValue');
-const getInside = require('../parsers/inside');
-const pipeSplit = require('../parsers/pipeSplit');
-const pipeList = require('../parsers/pipeList');
+const keyValue = require('./parsers/keyValue');
+const getInside = require('./parsers/inside');
+const pipeSplit = require('./parsers/pipeSplit');
+const pipeList = require('./parsers/pipeList');
 
 const parsers = {
